requirements.txt
PyPDF2
sentence-transformers
keybert
openai
transformers
torch
python-dotenv
pinecone